lib: Cast off_t to long long int
To fix warnings on MUSL 32bit:

../include/tst_safe_macros.h:284:50: warning: format '%ld' expects
argument of type 'long int', but argument 11 has type 'off_t' {aka 'long
long int'} [-Wformat=]
  284 |                 "mmap(%p, %zu, %s(%x), %d, %d, %ld)",
      |                                                ~~^
      |                                                  |
      |                                                  long int
      |                                                %lld
  285 |                 addr, length, prot_buf, prot, flags, fd, offset);
      |                                                          ~~~~~~
      |                                                          |
      |                                                          off_t {aka long long int}

Link: https://lore.kernel.org/ltp/20240523161944.GA29885@pevik/
Acked-by: Cyril Hrubis <chrubis@suse.cz>

+++ include/tst_safe_macros.h
@@ -281,8 +281,8 @@ static inline void *safe_mmap(const char *file, const int lineno,
 	tst_prot_to_str(prot, prot_buf);
 
 	tst_res_(file, lineno, TDEBUG,
-		"mmap(%p, %zu, %s(%x), %d, %d, %ld)",
-		addr, length, prot_buf, prot, flags, fd, offset);
+		"mmap(%p, %zu, %s(%x), %d, %d, %lld)",
+		addr, length, prot_buf, prot, flags, fd, (long long int)offset);
 
 	rval = mmap(addr, length, prot, flags, fd, offset);
 	if (rval == MAP_FAILED) {
